+++ test/acceptance/diffs.js
@@ -1,4 +1,3 @@
-
 var fs = require('fs')
   , cssin = fs.readFileSync('test/acceptance/fixtures/css.in', 'ascii')
   , cssout = fs.readFileSync('test/acceptance/fixtures/css.out', 'ascii');
@@ -83,5 +82,12 @@ describe('diffs', function(){
     var obj2 = { a: 123, b: 456 };
 
     // obj1.should.equal(obj2);
+  });
+
+  it('should display diff by data and not like an objects', function(){
+    var buf1 = new Buffer([0x01]);
+    var buf2 = new Buffer([0x02]);
+
+//    buf1.should.equal(buf2);
   });
 });