+++ test/acceptance/glob/glob.js
@@ -1,4 +1,3 @@
-
 describe('globbing test', function(){
   it('should find this test', function(){
     // see glob.sh for details