+++ test/browser/large.js
@@ -1,4 +1,3 @@
-
 var n = 30;
 while (n--) {
   describe('Array ' + n, function(){