+++ test/hook.sync.js
@@ -1,4 +1,3 @@
-
 describe('serial', function(){
   var calls = [];