+++ test/http.meta.2.js
@@ -1,4 +1,3 @@
-
 var http = require('http');
 
 var server = http.createServer(function(req, res){
@@ -53,6 +52,7 @@ function get(url) {
 
         expected = body;
         describe('GET ' + url, function(){
+          this.timeout(500);
           if (fields) {
             describe('when given ' + fields, function(){
               it('should respond with "' + body + '"', request);